Add files via upload
This is the demo of camera movement which is being controlled by the accelerometer mounted on VR headset to mimick the user's head movement. Accelerometer transfers the data through Bluetooth module to the Arduino which in return drives the relays to control the DC gear motors from which the camera is moved.